added new diagnostics, deleted redundant old diagnostics and moved files into more sensible locations
