+++ coppafish/call_spots/__init__.py
@@ -1,6 +1,6 @@
 from .qual_check import get_spot_intensity
 from .background import fit_background
-from .base import get_bled_codes, get_non_duplicate, compute_gene_efficiency
+from .base import get_non_duplicate, bayes_mean, compute_bleed_matrix
 from .qual_check import omp_spot_score, quality_threshold, get_intensity_thresh
 from .dot_product import dot_product_score, gene_prob_score
