+++ coppafish/plot/__init__.py
@@ -1,9 +1,5 @@
 from .results_viewer.base import Viewer
-from .call_spots import view_spot, view_codes, view_bled_codes, view_bleed_matrix, view_intensity, gene_counts
-from .call_spots import view_scaled_k_means
-from .call_spots.dot_product import view_score
-from .call_spots.weight import view_weight
-from .call_spots.background import view_background
+from .call_spots import view_spot, view_codes, view_bled_codes, view_bleed_matrix
 from .raw import view_raw
 from .extract import view_extracted_images
 from .filter import view_filtered_images
