+++ coppafish/plot/call_spots/__init__.py
@@ -1,4 +1,2 @@
-from .spot_colors import view_spot, view_codes, view_intensity
+from .spot_colors import view_spot, view_codes, GESpotViewer, GEViewer, BGNormViewer
 from .bleed_matrix import view_bleed_matrix, view_bled_codes, ColorPlotBase
-from .gene_counts import gene_counts
-from .scaled_k_means import view_scaled_k_means
