+++ coppafish/plot/call_spots/spot_colors.py
@@ -1,21 +1,28 @@
-import matplotlib.pyplot as plt
 import matplotlib as mpl
-import numpy as np
-from matplotlib.widgets import Button, RangeSlider
-from ...call_spots.qual_check import omp_spot_score, get_intensity_thresh
-from ...setup import Notebook
-from ...spot_colors import base
-import matplotlib
 from typing import List, Optional, Tuple, Union
+import matplotlib.colors as colors
+from matplotlib.patches import Rectangle
+import matplotlib.pyplot as plt
+from matplotlib.widgets import Button, Slider, RangeSlider
+import mplcursors
+import numpy as np
 
-plt.style.use("dark_background")
+from ...call_spots.qual_check import omp_spot_score
+from ...spot_colors import base
+from ...setup import Notebook
+
+try:
+    import importlib_resources
+except ModuleNotFoundError:
+    import importlib.resources as importlib_resources
 try:
     # So matplotlib plots pop out
     # Put in try so don't get error when unit testing in continuous integration
     # which is in headless mode
-    matplotlib.use("qtagg")
+    mpl.use("qtagg")
 except ImportError:
     pass
+plt.style.use("dark_background")
 
 
 class ColorPlotBase:
@@ -111,7 +118,7 @@ class ColorPlotBase:
             self.caxis_info[key]["min"] = -self.caxis_info[key]["max"]
             self.caxis_info[key]["clims"] = [self.caxis_info[key]["min"], self.caxis_info[key]["max"]]
             # cmap_norm is so cmap is white at 0.
-            self.caxis_info[key]["cmap_norm"] = matplotlib.colors.TwoSlopeNorm(
+            self.caxis_info[key]["cmap_norm"] = mpl.colors.TwoSlopeNorm(
                 vmin=self.caxis_info[key]["min"], vcenter=0, vmax=self.caxis_info[key]["max"]
             )
 
@@ -497,55 +504,469 @@ class view_spot(ColorPlotBase):
         plt.show()
 
 
-class view_intensity(ColorPlotBase):
-    def __init__(self, nb: Notebook, spot_no: int, method: str = "anchor"):
+# We are now going to create a new class that will allow us to view the spots used to calculate the gene efficiency
+# for a given gene. This will be useful for checking that the spots used are representative of the gene as a whole.
+class GEViewer:
+    def __init__(self, nb: Notebook):
         """
-        Diagnostic to show how intensity is computed from `spot_colour`.
-
+        Diagnostic to show the n_genes x n_rounds gene efficiency matrix as a heatmap.
         Args:
             nb: Notebook containing experiment details. Must have run at least as far as `call_reference_spots`.
-            spot_no: Spot of interest to be plotted.
-            method: `'anchor'` or `'omp'`.
-                Which method of gene assignment used i.e. `spot_no` belongs to `ref_spots` or `omp` page of Notebook.
         """
-        if method.lower() == "omp":
-            page_name = "omp"
-            t = nb.omp.tile[spot_no]
-            config = nb.get_config()["thresholds"]
-        else:
-            page_name = "ref_spots"
-            t = nb.ref_spots.tile[spot_no]
-        colour_norm = nb.call_spots.colour_norm_factor[t].transpose()
-        intensity_saved = nb.__getattribute__(page_name).intensity[spot_no]
-        intensity_thresh = get_intensity_thresh(nb)
-        spot_colour = (
-            nb.__getattribute__(page_name)
-            .colours[spot_no].transpose() * colour_norm
-        )
-        subplot_adjust = [0.07, 0.775, 0.1, 0.91]
-        super().__init__([spot_colour], colour_norm, subplot_adjust=subplot_adjust)
-        if intensity_saved > intensity_thresh:
-            colour = "w"
-        else:
-            colour = "r"
-        spot_colour_symbol = r"$\mathbf{\zeta_s}$"
-        intensity_symbol = r"$\chi_s$, (median of $\max_c\zeta_{s_{rc}}$ indicated in green)"
-        self.ax[0].set_title(
-            f"Spot colour, {spot_colour_symbol}, for spot {spot_no}\n"
-            f"Intensity, {intensity_symbol} = {str(np.around(intensity_saved, 3))}",
-            color=colour,
-        )
-        self.ax[0].set_yticks(ticks=np.arange(self.im_data[0].shape[0]), labels=nb.basic_info.use_channels)
-        self.ax[0].set_xticks(ticks=np.arange(self.im_data[0].shape[1]), labels=nb.basic_info.use_rounds)
-        self.ax[0].set_xlabel("Round")
-        self.fig.supylabel("colour Channel")
-        # Highlight max channel in each round which contributes to intensity
-        max_channels = np.argmax(self.im_data[0], axis=0)
-        for r in range(len(nb.basic_info.use_rounds)):
-            # can't add rectangle to multiple axes hence second for loop
-            rectangle = plt.Rectangle(
-                (r - 0.5, max_channels[r] - 0.5), 1, 1, fill=False, ec="lime", linestyle=":", lw=4
-            )
-            self.ax[0].add_patch(rectangle)
-        self.change_norm()  # initialise with method = 'norm'
+        self.nb = nb
+        self.n_genes = nb.call_spots.gene_efficiency.shape[0]
+        self.fig, self.ax = plt.subplots(figsize=(10, 10))
+        gene_efficiency = nb.call_spots.gene_efficiency
+        self.ax.imshow(gene_efficiency, cmap="viridis", vmin=0, vmax=gene_efficiency.max(), aspect="auto")
+        self.ax.set_xlabel("Round")
+        self.ax.set_ylabel("Gene")
+        self.ax.set_xticks(ticks=np.arange(gene_efficiency.shape[1]))
+        self.ax.set_yticks([])
+
+        # add colorbar
+        self.ax.set_title("Gene Efficiency")
+        cax = self.fig.add_axes([0.95, 0.1, 0.03, 0.8])
+        cbar = self.fig.colorbar(self.ax.images[0], cax=cax)
+        cbar.set_label("Gene Efficiency")
+
+        # Adding gene names to y-axis would be too crowded. We will use mplcursors to show gene name of gene[r] when
+        # hovering over row r of the heatmap. This means we need to only extract the y position of the mouse cursor.
+        gene_names = nb.call_spots.gene_names
+        mplcursors.cursor(self.ax, hover=True).connect("add", lambda sel: self.plot_gene_name(sel.index[0]))
+        # 2. Allow genes to be selected by clicking on them
+        mplcursors.cursor(self.ax, hover=False).connect("add", lambda sel: GESpotViewer(nb, sel.index[0]))
+        # 3. We would like to add a white rectangle around the observed spot when we hover over it. We will
+        # use mplcursors to do this. We need to add a rectangle to the plot when hovering over a gene.
+        # We also want to turn off annotation when hovering over a gene so we will use the `hover=False` option.
+        mplcursors.cursor(self.ax, hover=2).connect("add", lambda sel: self.add_rectangle(sel.index[0]))
+
         plt.show()
+
+    def add_rectangle(self, index):
+        # We need to remove any existing rectangles from the plot
+        index = max(0, index)
+        index = min(index, self.n_genes - 1)
+        for rectangle in self.ax.patches:
+            rectangle.remove()
+        # We can then add a new rectangle to the plot
+        self.ax.add_patch(Rectangle((-0.5, index - 0.5), self.nb.basic_info.n_rounds, 1, fill=False, edgecolor="white"))
+
+    def plot_gene_name(self, index):
+        # We need to remove any existing gene names from the plot
+        index = max(0, index)
+        index = min(index, self.n_genes - 1)
+        for text in self.ax.texts:
+            text.remove()
+        # We can then add a new gene name to the top right of the plot in size 20 font
+        self.ax.text(
+            1.05,
+            1.05,
+            self.nb.call_spots.gene_names[index],
+            transform=self.ax.transAxes,
+            size=20,
+            horizontalalignment="right",
+            verticalalignment="top",
+            color="white",
+        )
+
+
+class GESpotViewer:
+    def __init__(self, nb: Notebook, gene_index: int = 0, use_ge=True):
+        """
+        Diagnostic to show the spots used to calculate the gene efficiency for a given gene.
+        Args:
+            nb: Notebook containing experiment details. Must have run at least as far as `call_reference_spots`.
+            iteration: iteration of call spots that we would like to view
+            gene_index: Index of gene to be plotted.
+        """
+        self.nb = nb
+        self.gene_index = gene_index
+        self.n_genes = nb.call_spots.gene_efficiency.shape[0]
+        self.mode = "C"
+        self.use_ge = use_ge
+        # Load spots
+        self.load_spots(gene_index)
+        # Now initialise the plot, adding fig and ax attributes to the class
+        self.plot()
+
+        plt.show()
+
+    def load_spots(self, gene_index: int, use_ge=True):
+        nb = self.nb
+        n_channels = len(nb.basic_info.use_channels)
+        # First we need to find the spots used to calculate the gene efficiency for the given gene.
+        initial_assignment = np.argmax(nb.ref_spots.gene_probabilities, axis=1)
+        if use_ge:
+            self.gene_g_mask = nb.call_spots.use_ge * (initial_assignment == gene_index)
+        else:
+            self.gene_g_mask = nb.ref_spots.gene_no == gene_index
+        # self.gene_g_mask = nb.ref_spots.gene_no == gene_index
+        spots = nb.ref_spots.colours[self.gene_g_mask][:, :, nb.basic_info.use_channels]
+        # remove background codes. To do this, repeat background_strenth along a new axis for rounds
+        background_strength = nb.ref_spots.background_strength[self.gene_g_mask]
+        # remove background from spots
+        spots = spots - background_strength
+        spots = spots.reshape((spots.shape[0], spots.shape[1] * spots.shape[2]))
+        colour_norm = np.repeat(
+            np.mean(nb.call_spots.colour_norm_factor, axis=0)[
+                np.ix_(nb.basic_info.use_rounds, nb.basic_info.use_channels)
+            ].reshape((1, -1)),
+            spots.shape[0],
+            axis=0,
+        )
+        self.spots = spots * colour_norm
+        # order spots by nb.ref_spots.scores
+        # self.spots = self.spots[np.argsort(nb.ref_spots.scores[self.gene_g_mask]), :]
+        # We need to find the expected spot profile for each round/channel
+        self.spots_expected = nb.call_spots.bled_codes[self.gene_index, :, nb.basic_info.use_channels].T
+        # normalise so that mean brightness is the same as that of the observed spots
+        mean_row_strength = np.mean(np.sum(self.spots, axis=1))
+        self.spots_expected = self.spots_expected.reshape((1, -1)) * mean_row_strength / np.sum(self.spots_expected)
+        self.spots_expected = np.repeat(self.spots_expected, self.spots.shape[0], axis=0)
+        # Now for each spot we would like to store a dye_efficiency vector. This is the least squares solution to
+        # the equation spots[i, r, :] = dye_efficiency * spots_expected[0, r, :].
+        auxilliary_spots = self.spots.reshape((self.spots.shape[0], self.spots.shape[1] // n_channels, n_channels))
+        auxilliary_spots_expected = self.spots_expected.reshape(
+            (self.spots.shape[0], self.spots.shape[1] // n_channels, n_channels)
+        )[0, :, :]
+        self.dye_efficiency = np.zeros((auxilliary_spots.shape[0], auxilliary_spots.shape[1]))
+
+        for s in range(auxilliary_spots.shape[0]):
+            for r in range(auxilliary_spots.shape[1]):
+                a = auxilliary_spots[s, r]
+                b = auxilliary_spots_expected[r]
+                self.dye_efficiency[s, r] = np.dot(a, b) / np.dot(b, b)
+
+        self.dye_efficiency_norm = self.dye_efficiency / np.linalg.norm(self.dye_efficiency, axis=1)[:, np.newaxis]
+        # Estimate parameters of VMF distribution for dye efficiency norm
+        r_bar = np.linalg.norm(np.mean(self.dye_efficiency_norm, axis=0))
+        self.mu_hat = np.mean(self.dye_efficiency_norm, axis=0) / r_bar
+        self.kappa_hat = r_bar * (nb.basic_info.n_rounds - r_bar**2) / (1 - r_bar**2)
+        self.likelihood = np.sum(
+            np.repeat(self.mu_hat[np.newaxis, :], self.spots.shape[0], axis=0) * self.dye_efficiency_norm, axis=1
+        )
+
+        # order spots by likelihood of dye efficiency
+        self.spots = self.spots[np.argsort(self.likelihood), :]
+        self.spot_id = np.where(self.gene_g_mask)[0][np.argsort(self.likelihood)]
+
+    def plot_ge_cmap(self):
+        if not hasattr(self, "fig"):
+            self.fig, self.ax = plt.subplots(1, 2, figsize=(10, 10))
+        else:
+            del self.fig, self.ax
+            self.fig, self.ax = plt.subplots(1, 2, figsize=(10, 10))
+        # Now we can plot the spots. We want to create 2 subplots. One with the spots observed and one with the expected
+        # spots. We will use the same color scale for both subplots.
+        vmax = np.max([np.percentile(self.spots, 99), np.percentile(self.spots_expected, 99)])
+        vmin = np.min([np.percentile(self.spots, 1), np.percentile(self.spots_expected, 1)])
+        # We can then plot the spots observed and the spots expected.
+        self.ax[0].imshow(self.spots, cmap="viridis", vmin=vmin, vmax=vmax, aspect="auto", interpolation="none")
+        self.ax[1].imshow(
+            self.spots_expected, cmap="viridis", vmin=vmin, vmax=vmax, aspect="auto", interpolation="none"
+        )
+        # We can then add titles and axis labels to the subplots.
+        self.ax[0].set_ylim(0, self.spots.shape[0])
+        self.ax[1].set_ylim(0, self.spots.shape[0])
+        self.ax[0].set_xlim(-0.5, self.spots.shape[1] - 0.5)
+        self.ax[1].set_xlim(-0.5, self.spots.shape[1] - 0.5)
+        self.ax[0].set_title("Observed Spot Colours")
+        self.ax[1].set_title("Expected Spot Colours")
+        self.ax[0].set_xlabel("Spot Colour (flattened)")
+        self.ax[1].set_xlabel("Spot Colour (flattened)")
+        self.ax[0].set_ylabel("Spot")
+        self.ax[1].set_ylabel("Spot")
+
+        # We would like to add red vertical lines to show the start of each round.
+        for j in range(2):
+            for i in range(self.nb.basic_info.n_rounds):
+                self.ax[j].axvline(i * len(self.nb.basic_info.use_channels) - 0.5, color="r")
+
+        # Set supertitle, colorbar and show plot
+        self.fig.suptitle(
+            "Gene Efficiency Calculation for Gene "
+            + self.nb.call_spots.gene_names[self.gene_index]
+            + ". \n Estimated concentration parameter = "
+            + str(np.round(self.kappa_hat, 2))
+        )
+
+        # Add gene efficiency plot on top of prediction on the right
+        ge = self.nb.call_spots.gene_efficiency[self.gene_index]
+        ge_max = np.max(self.nb.call_spots.gene_efficiency)
+        ge = ge * self.spots.shape[0] / ge_max
+        ge = np.repeat(ge, self.nb.basic_info.n_rounds)
+        # Now clip the gene efficiency to be between 0 and the number of spots
+        ge = np.clip(ge, 0, self.spots.shape[0] - 1)
+        # We can then plot the gene efficiency
+        self.ax[1].plot(ge, color="w")
+        # plot a white line at the gene efficiency of 1.
+        self.ax[1].axhline(0, color="w", linestyle="--", label="Gene Efficiency = " + str(np.round(ge_max, 2)))
+        self.ax[1].axhline(self.spots.shape[0] / ge_max, color="w", linestyle="--", label="Gene Efficiency = 1")
+        self.ax[1].axhline(self.spots.shape[0] - 1, color="w", linestyle="--", label="Gene Efficiency = 0")
+        self.ax[1].legend(loc="upper right")
+
+        # For each round, we'd like to plot the dye efficiencies associated with the spots in this gene and round.
+        # First, for each round, sort dye_efficiencies in ascending order, then clip between 0 and ge_max.
+        for r in range(self.nb.basic_info.n_rounds):
+            dye_efficiency = self.dye_efficiency[:, r]
+            dye_efficiency = np.sort(dye_efficiency)
+            dye_efficiency = np.clip(dye_efficiency, 0, ge_max)
+            # Now we can plot the dye efficiencies. We want to plot them on the left subplot, and we want each round's
+            # dye efficiencies to be in between the red lines corresponding to that round.
+            # In order to make the dye_efficiencies appear in the correct place, we need to scale and shift them like
+            # we did with gene efficiency.
+            dye_efficiency = dye_efficiency * self.spots.shape[0] / ge_max
+            x_start, x_end = (
+                r * len(self.nb.basic_info.use_channels) - 0.5,
+                (r + 1) * len(self.nb.basic_info.use_channels) - 0.5,
+            )
+            self.ax[0].plot(np.linspace(x_start, x_end, len(dye_efficiency)), dye_efficiency, color="w")
+
+        # Add simple colorbar. Move this a little up to make space for the button.
+        cax = self.fig.add_axes([0.925, 0.1, 0.03, 0.7])
+        self.fig.colorbar(
+            plt.cm.ScalarMappable(norm=colors.Normalize(vmin=vmin, vmax=vmax), cmap="viridis"),
+            cax=cax,
+            label="Spot Colour",
+        )
+        self.add_cmap_widgets()
+        self.fig.canvas.draw_idle()
+
+    def plot_ge_hist(self, percentile=95):
+        """
+        Plot n_rounds histograms of all gene_efficiencies for each spot in each round.
+        """
+        nb = self.nb
+        gene_code = nb.call_spots.gene_codes[self.gene_index]
+
+        if not hasattr(self, "fig"):
+            # Create figure and axis
+            self.fig, self.ax = plt.subplots(1, nb.basic_info.n_rounds, figsize=(20, 5))
+        else:
+            del self.fig, self.ax
+            # Create figure and axis
+            self.fig, self.ax = plt.subplots(1, nb.basic_info.n_rounds, figsize=(20, 5))
+
+        fig, ax = self.fig, self.ax
+        # Loop through each round and plot the histogram of dye efficiencies
+        for r in range(self.nb.basic_info.n_rounds):
+            # Take max efficiency as percentile of rc_spot_intensity
+            max_efficiency = np.percentile(self.dye_efficiency_norm, percentile)
+            # Now plot both spot intensities on the same histogram
+            ax[r].hist(self.dye_efficiency_norm[:, r], bins=np.linspace(0, max_efficiency, 20), density=True)
+
+            # Add a box in the top right of the plot with the Gene Efficiency
+            ge = nb.call_spots.gene_efficiency[self.gene_index, r]
+            ax[r].set_yticks([])
+            ax[r].set_xticks([])
+            ax[r].set_xlabel("Relative Dye Efficiency \n Gene Efficiency = " + str(np.round(ge, 2)))
+            ax[r].set_ylabel("Frequency")
+
+        # Adjust subplots to leave space on the right for the legend and buttons
+        fig.subplots_adjust(right=0.9)
+
+        # Add a single legend on the top right of the figure. Want to take only first 2 labels from ax 0
+        legend_ax = fig.add_axes([0.925, 0.9, 0.05, 0.05])
+        legend_ax.axis("off")
+        fig.suptitle("Histogram of relative dye efficiencies for Gene " + nb.call_spots.gene_names[self.gene_index])
+        self.add_hist_widgets()
+        fig.canvas.draw_idle()
+
+    def plot(self):
+        # Get rid of any existing plots
+        if hasattr(self, "fig"):
+            plt.close(self.fig)
+
+        # Plot the correct plot
+        if self.mode == "C":
+            self.plot_ge_cmap()
+        elif self.mode == "H":
+            self.plot_ge_hist()
+
+        self.add_switch_buttons()
+
+    def button_cmap_clicked(self, event):
+        self.mode = "C"
+        self.plot()
+
+    def button_all_clicked(self, event):
+        self.use_ge = not self.use_ge
+        self.load_spots(self.gene_index, self.use_ge)
+        self.plot()
+
+    def button_hist_clicked(self, event):
+        self.mode = "H"
+        self.plot()
+
+    def add_switch_buttons(self):
+        # Add 2 buttons on the bottom right of the figure allowing the user to choose between viewing the
+        # histogram or the colourmap. Make text black so that it is visible on the white background.
+        ax_button = self.fig.add_axes([0.925, 0.05, 0.02, 0.03])
+        self.button_hist = Button(ax_button, "H", color="black")
+        self.button_hist.on_clicked(self.button_hist_clicked)
+        ax_button_cmap = self.fig.add_axes([0.95, 0.05, 0.02, 0.03])
+        self.button_cmap = Button(ax_button_cmap, "C", color="black")
+        self.button_cmap.on_clicked(self.button_cmap_clicked)
+        # Add button to show all spots in the gene. Put this in the top right of the figure
+        ax_button_show_all = self.fig.add_axes([0.925, 0.825, 0.05, 0.05])
+        self.button_all = Button(ax_button_show_all, "ALL", color="black")
+        self.button_all.on_clicked(self.button_all_clicked)
+
+    def add_cmap_widgets(self):
+        # Initialise buttons and cursors
+        # 1. We would like each row of the plot to be clickable, so that we can view the observed spot.
+        mplcursors.cursor(self.ax[0], hover=False).connect(
+            "add", lambda sel: view_codes(self.nb, self.spot_id[sel.index[0]])
+        )
+        # 2. We would like to add a white rectangle around the observed spot when we hover over it
+        mplcursors.cursor(self.ax[0], hover=2).connect("add", lambda sel: self.add_rectangle(sel.index[0]))
+
+    def add_hist_widgets(self):
+        # Add a slider on the right of the figure allowing the user to choose the percentile of the histogram
+        # to use as the maximum intensity. This slider should be the same dimensions as the colorbar and should
+        # be in the same position as the colorbar. We should slide vertically to change the percentile.
+        self.ax_slider = self.fig.add_axes([0.94, 0.15, 0.02, 0.6])
+        self.slider = Slider(self.ax_slider, "Percentile", 80, 100, valinit=90, orientation="vertical")
+        self.slider.on_changed(lambda val: self.update_hist(int(val)))
+
+    def update_hist(self, percentile):
+        nb = self.nb
+        fig, ax = self.fig, self.ax
+        # Loop through each round and plot the histogram of dye efficiencies
+        for r in range(self.nb.basic_info.n_rounds):
+            # Take max efficiency as percentile of rc_spot_intensity
+            max_efficiency = np.percentile(self.dye_efficiency_norm, percentile)
+            # Now plot both spot intensities on the same histogram
+            ax[r].cla()
+            ax[r].hist(self.dye_efficiency_norm[:, r], bins=np.linspace(0, max_efficiency, 20), density=True)
+
+            # Add a box in the top right of the plot with the Gene Efficiency
+            ge = nb.call_spots.gene_efficiency[self.gene_index, r]
+            ax[r].set_yticks([])
+            ax[r].set_xticks([])
+            ax[r].set_xlabel("Spot Intensity \n Gene Efficiency = " + str(np.round(ge, 2)))
+            ax[r].set_ylabel("Frequency")
+
+        # Adjust subplots to leave space on the right for the legend and buttons
+        fig.subplots_adjust(right=0.9)
+        fig.canvas.draw_idle()
+
+    def add_rectangle(self, index):
+        # We need to remove any existing rectangles from the plot
+        index = max(0, index)
+        index = min(index, self.spots.shape[0] - 1)
+        for rectangle in self.ax[0].patches:
+            rectangle.remove()
+        # We can then add a new rectangle to the plot
+        self.ax[0].add_patch(
+            Rectangle(
+                (-0.5, index - 0.5),
+                self.nb.basic_info.n_rounds * len(self.nb.basic_info.use_channels),
+                1,
+                fill=False,
+                edgecolor="white",
+            )
+        )
+
+
+class BGNormViewer:
+    """
+    This function will plot all spots before and after background subtraction and order them by background noise.
+    We will then plot the normalised spots too.
+    """
+
+    def __init__(self, nb):
+        self.nb = nb
+        isolated = nb.ref_spots.isolated
+        tile = nb.ref_spots.tile
+        n_spots = np.sum(isolated)
+        n_rounds, n_channels_use = len(nb.basic_info.use_rounds), len(nb.basic_info.use_channels)
+        norm_factor = nb.call_spots.colour_norm_factor[tile][isolated]
+        background_noise = nb.ref_spots.background_strength[isolated]
+
+        spot_colour_raw = nb.ref_spots.colours.copy()[isolated][:, :, nb.basic_info.use_channels]
+        spot_colour_no_bg = spot_colour_raw - background_noise
+        spot_colour_normed_no_bg = spot_colour_no_bg * norm_factor[isolated]
+        # Now we'd like to order the spots by background noise in descending order
+        background_noise = np.sum(abs(background_noise), axis=(1, 2))
+        spot_colour_raw = spot_colour_raw[np.argsort(background_noise)[::-1]]
+        spot_colour_no_bg = spot_colour_no_bg[np.argsort(background_noise)[::-1]]
+        spot_colour_normed_no_bg = spot_colour_normed_no_bg[np.argsort(background_noise)[::-1]]
+        # Finally, we need to reshape the spots to be n_spots x n_rounds * n_channels. Since we want the channels to be
+        # in consecutive blocks of size n_rounds, we can reshape by first switching the round and channel axes.
+        spot_colour_raw = spot_colour_raw.transpose((0, 2, 1))
+        spot_colour_raw = spot_colour_raw.reshape((n_spots, n_rounds * n_channels_use))
+        spot_colour_no_bg = spot_colour_no_bg.transpose((0, 2, 1))
+        spot_colour_no_bg = spot_colour_no_bg.reshape((n_spots, n_rounds * n_channels_use))
+        spot_colour_normed_no_bg = spot_colour_normed_no_bg.transpose((0, 2, 1))
+        spot_colour_normed_no_bg = spot_colour_normed_no_bg.reshape((n_spots, n_rounds * n_channels_use))
+
+        # We're going to make a little viewer to show spots before and after background subtraction
+        colour_scaling_factor = 10
+        fig, ax = plt.subplots(1, 3, figsize=(10, 5))
+        max_intensity = np.max(spot_colour_raw)
+        min_intensity = np.min(spot_colour_raw)
+        ax[0].imshow(
+            spot_colour_raw,
+            aspect="auto",
+            vmin=min_intensity * colour_scaling_factor,
+            vmax=max_intensity * colour_scaling_factor,
+            interpolation="none",
+        )
+        ax[0].set_title("Before background subtraction and normalisation")
+        ax[0].set_xticks([])
+
+        max_intensity = np.max(spot_colour_no_bg)
+        min_intensity = np.min(spot_colour_no_bg)
+        ax[1].imshow(
+            spot_colour_no_bg,
+            aspect="auto",
+            vmin=min_intensity * colour_scaling_factor,
+            vmax=max_intensity * colour_scaling_factor,
+            interpolation="none",
+        )
+        ax[1].set_title("After bg removal, before normalisation")
+        ax[1].set_xticks([])
+
+        max_intensity = np.max(spot_colour_normed_no_bg)
+        min_intensity = np.min(spot_colour_normed_no_bg)
+        ax[2].imshow(
+            spot_colour_normed_no_bg,
+            aspect="auto",
+            vmin=min_intensity * colour_scaling_factor,
+            vmax=max_intensity * colour_scaling_factor,
+            interpolation="none",
+        )
+        ax[2].set_title("After background subtraction and normalisation")
+        ax[2].set_xticks([])
+
+        # Now add vertical dashed red lines to separate channels
+        for i in range(n_rounds - 1):
+            for j in range(3):
+                ax[j].axvline((i + 1) * n_channels_use - 0.5, color="r", linestyle="--")
+
+        # For each ax, add the channels to the x axis. Since the x axis has n_rounds channels, followed by n_rounds
+        # channels, etc, we only need to add the channels once. We want to add the channel name at the bottom of
+        # each set of channels, so we'll add the channel name at the position of the middle channel in each set
+        for i in range(n_channels_use):
+            y_pos = int(n_spots * (1 + 0.05))
+            for j in range(3):
+                ax[j].text(i * n_rounds + n_rounds // 2, y_pos, nb.basic_info.use_channels[i], color="w", fontsize=10)
+
+        # Add a title
+        fig.suptitle("Background subtraction and normalisation", fontsize=16)
+
+        plt.show()
+
+    # add slider to allow us to vary value of interp between 0 and 1 and update plot
+    # def add_hist_widgets(self):
+    #     Add a slider on the right of the figure allowing the user to choose the percentile of the histogram
+    #     to use as the maximum intensity. This slider should be the same dimensions as the colorbar and should
+    #     be in the same position as the colorbar. We should slide vertically to change the percentile.
+    # self.ax_slider = self.fig.add_axes([0.94, 0.15, 0.02, 0.6])
+    # self.slider = Slider(self.ax_slider, 'Interpolation Coefficient', 0, 1, valinit=0, orientation='vertical')
+    # self.slider.on_changed(lambda val: self.update_hist(int(val)))
+    #
+    # TODO: Add 2 buttons, one for separating normalisation by channel and one for separating by round and channel
+
+
