--- coppafish/plot/call_spots_new/__init__.py
@@ -1,1 +1,0 @@
-from .base import GEViewer, ViewAllGeneScores, ViewBleedCalc, BGNormViewer
