+++ coppafish/plot/omp/__init__.py
@@ -1,2 +1,1 @@
 from .coefs import ViewOMPImage, ViewOMPPixelCoefficients
-from .score_hist import histogram_score, histogram_2d_score
